pyproject.toml
[tool.poetry]
name = "har-encoder-trainer"
version = "0.1.0"
description = "Human Activity Recognition Encoder Training with Self-Supervised Learning"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "encoder_training"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
torch = ">=2.1.0,<2.7.0"
matplotlib = ">=3.7.2"
numpy = ">=1.25.2"
scipy = ">=1.11.2"
scikit-learn = ">=1.3.0"
pandas = ">=2.0.3"
pyyaml = ">=6.0"
tqdm = ">=4.65.0"
joblib = ">=1.3.2"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.0.0"
black = ">=23.1.0"
isort = ">=5.12.0"
flake8 = ">=6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 100

[tool.flake8]
max-line-length = 100
exclude = [".git", "__pycache__", "build", "dist"] 